ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
onnxruntime-gpu==1.20.1
insightface==0.7.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==11.0.0
pydantic==2.10.3
python-multipart==0.0.18
httpx==0.28.1
redis==5.2.1
